--- handl_buf.c
@@ -1,21 +1,0 @@
-#include "main.h"
-
-/**
- * handl_buf - concatenates buffer chars
- * @buf: buffer pointer
- * @c: charcter to concatenate
- * @ibuf: index of buffer pointer
- * Return: index of buffer pointer
- */
-
-unsigned int handl_buf(char *buf, char c, unsigned int ibuf)
-{
-	if (ibuf == 1024)
-	{
-		print_buf(buf, ibuf);
-		ibuf = 0;
-	}
-	buf[ibuf] = c;
-	ibuf++;
-	return (ibuf);
-}
