--- print_chr.c
@@ -1,21 +1,0 @@
-#include "main.h"
-#include <stdio.h>
-
-/**
- * print_chr - write character c to stdout
- * @arguments: input char
- * @buf: buffer pointer
- * @ibuf: index for buffer pointer
- * Return: On success 1
- */
-
-
-int print_chr(va_list arguments, char *buf, unsigned int ibuf)
-{
-	char c;
-
-	c = va_arg(arguments, int);
-	handl_buf(buf, c, ibuf);
-
-	return (1);
-}
