+++ print_functions.c
@@ -1,0 +1,121 @@
+#include "main.h"
+#include <stdio.h>
+
+/**
+ * print_chr - write character c to stdout
+ * @arguments: input char
+ * @buf: buffer pointer
+ * @ibuf: index for buffer pointer
+ * Return: On success 1
+ */
+
+
+int print_chr(va_list arguments, char *buf, unsigned int ibuf)
+{
+	char c;
+
+	c = va_arg(arguments, int);
+	handl_buf(buf, c, ibuf);
+
+	return (1);
+}
+
+/**
+ * print_buf - prints buffer
+ * @buf: buffer pointer
+ * @nbuf: num of bytes to print
+ * Return: num of bytes printed
+ */
+
+int print_buf(char *buf, unsigned int nbuf)
+{
+	return (write(1, buf, nbuf));
+}
+
+/**
+ * print_int - prints an int
+ * @arguments: input string
+ * @buf: buffer pointer
+ * @ibuf: index for buffer pointer
+ * Return: number of chars printed
+ */
+
+int print_int(va_list arguments, char *buf, unsigned int ibuf)
+{
+	int int_input;
+	unsigned int int_in, int_temp, i, div, isneg;
+
+	int_input = va_arg(arguments, int);
+	isneg = 0;
+	if (int_input < 0)
+	{
+		int_in = int_input * -1;
+		ibuf = handl_buf(buf, '-', ibuf);
+		isneg = 1;
+	}
+	else
+	{
+		int_in = int_input;
+	}
+
+	int_temp = int_in;
+	div = 1;
+
+	while (int_temp > 9)
+	{
+		div *= 10;
+		int_temp /= 10;
+	}
+
+	for (i = 0; div > 0; div /= 10, i++)
+	{
+		ibuf = handl_buf(buf, ((int_in / div) % 10) + '0', ibuf);
+	}
+	return (i + isneg);
+}
+
+/**
+ * print_str - writes string to stdout
+ * @arguments: input string
+ * @buf: buffer pointer
+ * @ibuf: index for buffer pointer
+ * Return: On success 1
+ */
+
+int print_str(va_list arguments, char *buf, unsigned int ibuf)
+{
+	char *str;
+	unsigned int i;
+	char nill[] = "(null)";
+
+	str = va_arg(arguments, char*);
+	if (str == NULL)
+	{
+		for (i = 0; nill[i]; i++)
+			ibuf = handl_buf(buf, nill[i], ibuf);
+		return (6);
+	}
+	for (i = 0; str[i]; i++)
+		ibuf = handl_buf(buf, str[i], ibuf);
+	return (i);
+}
+
+/**
+ * handl_buf - concatenates buffer chars
+ * @buf: buffer pointer
+ * @c: charcter to concatenate
+ * @ibuf: index of buffer pointer
+ * Return: index of buffer pointer
+ */
+
+unsigned int handl_buf(char *buf, char c, unsigned int ibuf)
+{
+	if (ibuf == 1024)
+	{
+		print_buf(buf, ibuf);
+		ibuf = 0;
+	}
+	buf[ibuf] = c;
+	ibuf++;
+	return (ibuf);
+}
